refactor(plugin): delete initSocket and the socketio option
initSocket had zero import sites across crm, helpdesk, gameplan, builder,
insights, drive, suite and frappe's own ui/ package — every one of them
defines its own. What it did have was a plugin option defaulting to true, so
crm and helpdesk opened a second live socket.io connection on every page load
and then overwrote $socket with their own.

socket.io-client stays a dependency: resources/realtime.ts exports functions
typed against its Socket, and frappe/DataImport now constructs its own
connection — it was the last in-repo caller of initSocket.
